Fix long compare value datatype

+++ src/oehen.arguard.Tests/LessThan/ArgumentLongLessThanGuardTest.cs
@@ -14,7 +14,7 @@ namespace oehen.arguard
         [InlineData(5, 5)]
         public void
             ThrowIfIsLessOrEqualThanCompareValue_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsLessOrEqualThanTheCompareValue(
-                long argument, int compareValue)
+                long argument, long compareValue)
         {
             var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
             {
@@ -30,7 +30,7 @@ namespace oehen.arguard
         [InlineData(-3, 5)]
         public void
             ThrowIfIsLessThanCompareValue_ShouldThrowArgumentOutOfRangeException_IfArgumentValueIsLessThanTheCompareValue(
-                long argument, int compareValue)
+                long argument, long compareValue)
         {
             var exception = Assert.Throws<ArgumentOutOfRangeException>(() =>
             {
@@ -47,7 +47,7 @@ namespace oehen.arguard
         [InlineData(3, 3)]
         public void
             ThrowIfIsLessThanCompareValue_ShouldNotThrowArgumentOutOfRangeException_IfArgumentValueIsGreaterThanTheCompareValue(
-                long argument, int compareValue)
+                long argument, long compareValue)
         {
             var exception = Record.Exception(() => { argument.ThrowIfIsLessThan(compareValue, nameof(argument)); });
             Assert.Null(exception);
@@ -59,7 +59,7 @@ namespace oehen.arguard
         [InlineData(5, 3)]
         public void
             ThrowIfIsLessOrEqualThanCompareValue_ShouldNotThrowArgumentOutOfRangeException_IfArgumentValueIsGreaterOrEqualThanTheCompareValue(
-                long argument, int compareValue)
+                long argument, long compareValue)
         {
             var exception = Record.Exception(() =>
             {
